Move comptutation-engines outside of api folder
