+++ src/components/interpreter.tsx
@@ -11,10 +11,10 @@ import {
   refresh,
   stepForward,
   stepToIndex,
-} from "../api/computation-engines/manual/execute";
-import { extractManual } from "../api/computation-engines/manual/extract";
-import { isAtBlock } from "../api/computation-engines/manual/interpreter";
-import { Step } from "../api/computation-engines/manual/step";
+} from "../computation-engines/manual/execute";
+import { extractManual } from "../computation-engines/manual/extract";
+import { isAtBlock } from "../computation-engines/manual/interpreter";
+import { Step } from "../computation-engines/manual/step";
 import { executionStore as ctx } from "../api/execution";
 import { IEnvironment } from "../types/environment";
 import { extractViews } from "../util/acorn";
