+++ src/computation-engines/manual/constants.ts
