+++ src/computation-engines/singular-formula-solver.ts
